projects/nyuv2-semantic-segmentation/requirements.txt
matplotlib
numpy
pillow
torch
torchvision
tqdm
